Replace incorrect kline second timestamps instead of actual millisecond
It looks like those timestamps are not representing the real precision returned by the exchange. 
The actual data is 13 digits long which makes a big difference on how you want to handle it client-side as to decide if it should be divided or multiplied by 1000.

You guys might want to check the other endpoints but I can only attest for the klines as it is the one I use.